Move workflow activity to root package
