+++ docs/dapr-sdk-workflows/io/dapr/workflows/runtime/class-use/WorkflowActivity.html
@@ -1,8 +1,8 @@
 <!DOCTYPE HTML>
 <html lang="en">
 <head>
-<!-- Generated by javadoc (17) on Mon Dec 09 17:48:11 UTC 2024 -->
-<title>Uses of Interface io.dapr.workflows.runtime.WorkflowActivity (dapr-sdk-workflows 0.13.1 API)</title>
+<!-- Generated by javadoc (17) on Tue Aug 13 18:03:07 PDT 2024 -->
+<title>Uses of Interface io.dapr.workflows.runtime.WorkflowActivity (dapr-sdk-workflows 0.12.0 API)</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <meta name="dc.created" content="2024-12-09">
