+++ sdk-tests/src/test/java/io/dapr/it/testcontainers/FirstActivity.java
@@ -13,7 +13,7 @@ limitations under the License.
 
 package io.dapr.it.testcontainers;
 
-import io.dapr.workflows.runtime.WorkflowActivity;
+import io.dapr.workflows.WorkflowActivity;
 import io.dapr.workflows.runtime.WorkflowActivityContext;
 
 public class FirstActivity implements WorkflowActivity {
