+++ sdk-workflows/src/main/java/io/dapr/workflows/WorkflowActivity.java
@@ -11,7 +11,9 @@
 limitations under the License.
 */
 
-package io.dapr.workflows.runtime;
+package io.dapr.workflows;
+
+import io.dapr.workflows.runtime.WorkflowActivityContext;
 
 /**
  * Common interface for task activity implementations.
