+++ sdk-workflows/src/test/java/io/dapr/workflows/runtime/WorkflowRuntimeBuilderTest.java
@@ -14,6 +14,7 @@ package io.dapr.workflows.runtime;
 
 
 import io.dapr.workflows.Workflow;
+import io.dapr.workflows.WorkflowActivity;
 import io.dapr.workflows.WorkflowStub;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
